Update git-integration.md
From a recent ticket I worked on (#2990901), I understand we can benefit from adding to our docs that You can only have one GitHub Integration with only one Cloudflare account. Each GitHub account can be linked to one Cloudflare account, independent of which Cloudflare user/member is interacting with that account. So it would not be possible for example to link the same GitHub account to a Dev and a Prod Cloudflare account at the same time, but would be possible for multiple Cloudflare account members to interact with it once linked. 

PCX-9127